+++ noodles-util/src/variant/indexed_reader/builder.rs
@@ -5,6 +5,7 @@ use std::{
 };
 
 use noodles_bcf as bcf;
+use noodles_csi as csi;
 use noodles_vcf as vcf;
 
 use super::IndexedReader;
@@ -15,9 +16,16 @@ use crate::variant::{Compression, Format};
 pub struct Builder {
     compression: Option<Option<Compression>>,
     format: Option<Format>,
+    index: Option<csi::Index>,
 }
 
 impl Builder {
+    /// Sets an index.
+    pub fn set_index(mut self, index: csi::Index) -> Self {
+        self.index = Some(index);
+        self
+    }
+
     /// Builds an indexed variant reader from a path.
     pub fn build_from_path<P>(self, src: P) -> io::Result<IndexedReader<File>>
     where
@@ -38,12 +46,24 @@ impl Builder {
         };
 
         match (format, compression) {
-            (Format::Vcf, Some(Compression::Bgzf)) => vcf::indexed_reader::Builder::default()
-                .build_from_path(src)
-                .map(IndexedReader::Vcf),
-            (Format::Bcf, Some(Compression::Bgzf)) => bcf::indexed_reader::Builder::default()
-                .build_from_path(src)
-                .map(IndexedReader::Bcf),
+            (Format::Vcf, Some(Compression::Bgzf)) => {
+                let mut builder = vcf::indexed_reader::Builder::default();
+
+                if let Some(index) = self.index {
+                    builder = builder.set_index(index);
+                }
+
+                builder.build_from_path(src).map(IndexedReader::Vcf)
+            }
+            (Format::Bcf, Some(Compression::Bgzf)) => {
+                let mut builder = bcf::indexed_reader::Builder::default();
+
+                if let Some(index) = self.index {
+                    builder = builder.set_index(index);
+                }
+
+                builder.build_from_path(src).map(IndexedReader::Bcf)
+            }
             (_, None) => Err(io::Error::new(
                 io::ErrorKind::InvalidData,
                 "source not bgzip-compressed",